feat(__navigation.svelte): add preview alert message with cookie support
Add a preview alert message to the navigation section of the website when
PUBLIC_ENV is set to "preview". The message is hidden by default and can be
closed by clicking on the "X" icon. The message state is saved in a cookie
with a max age of one week.