feat(skills): add link property to all skills and add Google skill
Add a link property to all skills in the skills.ts file. The link property is a URL that points to the official website of the skill. Also, add a new skill, Google, to the list of skills. The Google skill is used to search for information using the Google Custom Search API. Finally, add the Google icon to the static/icons directory.