feat(projects.ts): add new project "Tweeets"
Add a new project called "Tweeets" to the list of projects. This project is a tweet generator with artificial intelligence, and it is free and open source. It uses the following technologies: Typescript, NextJs, OpenAI, Tailwindcss, Vercel, Prisma, Postgresql, Supabase, and Stripe. The project started on 06/08/2023. It has GitHub repository and live website links. The project type is "Open Source" and it has associated images. The color of the project is "#070709".